backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["jonxlegasa <jon.legasa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
config = "^0.5.1"
llama-index = "^0.10.42"
langchain-openai = "^0.1.8"
python-dotenv = "^1.0.1"
pydantic = "^2.7.3"
chromadb = "^0.5.0"
llama-index-vector-stores-chroma = "^0.1.8"
llama-index-embeddings-huggingface = "^0.2.1"
pygithub = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.poetry.scripts]
run-my-project = "backend.app:main"
